requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --upgrade-package djangorestframework-stubs -o requirements.txt requirements.in
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
    #   django-htmx
asttokens==2.4.1
    # via stack-data
attrs==24.2.0
    # via hypothesis
beautifulsoup4==4.12.3
billiard==4.2.1
    # via celery
boto3==1.35.39
botocore==1.35.39
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
decorator==5.1.1
    # via ipython
distro==1.9.0
    # via openai
dj-database-url==2.2.0
django==5.2.5
    # via
    #   dj-database-url
    #   django-cors-headers
    #   django-cotton
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-htmx
    #   django-redis
    #   django-storages
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-cors-headers==4.5.0
django-cotton==2.1.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-htmx==1.19.0
django-redis==5.4.0
django-storages==1.14.4
django-stubs==5.2.2
    # via djangorestframework-stubs
django-stubs-ext==5.2.2
    # via django-stubs
djangorestframework==3.16.1
djangorestframework-stubs==3.16.2
executing==2.1.0
    # via stack-data
factory-boy==3.3.1
faker==30.3.0
    # via factory-boy
gunicorn==23.0.0
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via openai
hypothesis==6.115.0
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
ipython==8.28.0
jedi==0.19.1
    # via ipython
jiter==0.6.1
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.4.2
    # via celery
lolwrapper @ git+https://github.com/brianjp93/lolapi.git@c84841786484f7e0b844fa203c06f8bdea54647f
matplotlib-inline==0.1.7
    # via ipython
mypy==1.13.0
    # via
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==1.0.0
    # via mypy
openai==1.51.2
packaging==24.1
    # via gunicorn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.4.0
prompt-toolkit==3.0.48
    # via
    #   click-repl
    #   ipython
psycopg==3.2.3
psycopg-binary==3.2.3
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.9.2
    # via openai
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via ipython
pytailwindcss==0.2.0
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   faker
python-decouple==3.8
pytz==2024.2
redis==4.6.0
    # via
    #   celery
    #   django-redis
requests==2.32.3
    # via
    #   djangorestframework-stubs
    #   lolwrapper
s3transfer==0.10.3
    # via boto3
sentry-sdk==2.16.0
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sortedcontainers==2.4.0
    # via hypothesis
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
tqdm==4.66.5
    # via openai
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
types-pyyaml==6.0.12.20240917
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.4.20250809
    # via djangorestframework-stubs
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   faker
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via
    #   celery
    #   kombu
urllib3==2.2.3
    # via
    #   botocore
    #   requests
    #   sentry-sdk
    #   types-requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.8.2
